fix: server перенести мидлвари и хендлеры в роутер
