+++ internal/server/router/handlers/json.get.go
@@ -7,21 +7,21 @@ import (
 	"github.com/mailru/easyjson"
 	"github.com/thefrol/kysh-kysh-meow/internal/metrica"
 	"github.com/thefrol/kysh-kysh-meow/internal/server/domain"
-	"github.com/thefrol/kysh-kysh-meow/internal/server/httpio"
+	"github.com/thefrol/kysh-kysh-meow/internal/server/router/httpio"
 )
 
-func (h *ForJSON) Update(w http.ResponseWriter, r *http.Request) {
+func (h *ForJSON) Get(w http.ResponseWriter, r *http.Request) {
 	var m metrica.Metrica
 
 	err := easyjson.UnmarshalFromReader(r.Body, &m)
 	if err != nil {
 		httpio.HTTPErrorWithLogging(w,
-			http.StatusBadRequest,
+			http.StatusInternalServerError,
 			"Не могу размаршалить в json тело запроса: %v", err)
 		return
 	}
 
-	v, err := h.Manager.UpdateMetrica(r.Context(), m)
+	v, err := h.Manager.GetMetrica(r.Context(), m)
 	if err != nil {
 		if errors.Is(err, domain.ErrorMetricNotFound) {
 			httpio.HTTPErrorWithLogging(w,
@@ -29,7 +29,6 @@ func (h *ForJSON) Update(w http.ResponseWriter, r *http.Request) {
 				"Не удалось найти метрику %v типа %v: %v", m.ID, m.MType, err)
 			return
 		}
-
 		httpio.HTTPErrorWithLogging(w,
 			http.StatusBadRequest,
 			"Получена неправильная метрика %v.%v: %v", m.MType, m.ID, err)
